Revert "Plan 9 from Bell Labs Fourth Edition 2005-11-07"
This reverts commit 4de34a7edde43207e841ec91ecd12e6cf5f5ebe7.

+++ sys/src/cmd/acid/print.c
@@ -109,7 +109,8 @@ whatis(Lsym *l)
 		if(t == TINT || t == TFLOAT)
 			Bprint(bout, " format %c", l->v->fmt);
 		if(l->v->comt)
-			Bprint(bout, " complex %s", l->v->comt->base->name);
+			Bprint(bout, " complex %s",
+						l->v->comt->base->name);
 		Bputc(bout, '\n');
 		def = 1;
 	}
@@ -251,7 +252,7 @@ pexpr(Node *n)
 	case OCONST:
 		switch(n->type) {
 		case TINT:
-			Bprint(bout, "%lld", n->ival);
+			Bprint(bout, "%d", (int)n->ival);
 			break;
 		case TFLOAT:
 			Bprint(bout, "%g", n->fval);
